TG-118: Components style fixes

+++ src/Components/Employees/index.js
@@ -3,7 +3,6 @@ import { useHistory } from 'react-router-dom';
 import { useSelector, useDispatch } from 'react-redux';
 import { deleteEmployee, getEmployees } from 'redux/Employees/thunks.js';
 import styles from './employees.module.css';
-import Button from 'Components/Shared/Button';
 import Table from 'Components/Shared/Table';
 import Modal from 'Components/Shared/Modal';
 import Loader from 'Components/Shared/Loader/index.js';
@@ -89,25 +88,10 @@ const Employees = () => {
         <Modal
           heading={`Are you sure you want to delete employee ${selectedEmployee.name} ${selectedEmployee.lastName}?`}
           setModalDisplay={setIsModal}
-          theme={'confirm'}
-        >
-          <p>This change cannot be undone!</p>
-          <Button
-            label={'Cancel'}
-            theme={'primary'}
-            onClick={() => {
-              setIsModal();
-            }}
-          />
-          <Button
-            label="Confirm"
-            theme="tertiary"
-            onClick={() => {
-              deleteItem();
-              setIsModal(false);
-            }}
-          />
-        </Modal>
+          theme="confirm"
+          message="This change cannot be undone!"
+          confirmFunction={deleteItem()}
+        />
       )}
     </section>
   );
